refactor: Board클래스를 기존의 ChessBoard클래스와 병합
